pocs/ecr_docker/app_base/requirements.txt
fastapi
uvicorn[standard]
jinja2
python-dotenv
openai==0.28
markdown
aiosqlite
itsdangerous
python-multipart
requests